+++ features/video-script-engine/actions/index.ts
@@ -1,0 +1,4 @@
+export * from "./create-video-script.action";
+export * from "./update-video-script.action";
+export * from "./delete-video-script.action";
+export * from "./generate-video-script.action";