+++ features/video-script-engine/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./VideoScriptCard";
+export * from "./VideoScriptList";
+export * from "./VideoScriptForm";